Split JDK8+ CompletableFuture tests to dedicated file (compiled only on Scala 3 becouse of build issues)
